requirements.txt
streamlit
streamlit-authenticator
openai
pandas
numpy
matplotlib
opencv-python
Pillow
pdf2image
tiktoken